fix(lib): clang-tidy redundant cast warnings
This patch removes redundant casts to the same type, as advised by the
clang-tidy google-readability-casting check.

+++ lib/attestation/src/attestation_token.c
@@ -331,5 +331,5 @@ int attest_realm_token_create(enum hash_algo algorithm,
 	QCBOREncode_CloseBstrWrap2(&(ctx->ctx.cbor_enc_ctx), false,
 				   &(ctx->ctx.signed_payload));
 
-	return (int)ATTEST_TOKEN_ERR_SUCCESS;
+	return ATTEST_TOKEN_ERR_SUCCESS;
 }
